Mem calculus are improved
git-svn-id: https://scm.gforge.inria.fr/svn/rtcelog/software/JBenchmarkerCE@1504 47090b3f-bbb5-42d7-8af4-003170c53eba

+++ src/main/java/crdt/simulator/sizecalculator/SizeCalculator.java
@@ -27,6 +27,12 @@ import java.io.IOException;
  * @author Stephane Martin <stephane@stephanemartin.fr>
  */
 public interface SizeCalculator {
-    public int serializ(CRDT m) throws IOException;
+    /**
+     *
+     * @param m the value of m
+     * @return the long
+     * @throws IOException
+     */
+    public long serializ(CRDT m) throws IOException;
     
 }
