+++ src/main/java/crdt/tree/orderedtree/renderer/SizeJSonDoc.java
@@ -45,8 +45,14 @@ public class SizeJSonDoc extends SizeXMLDoc {
         return "";
     }
 
+    /**
+     *
+     * @param m the value of m
+     * @return the long
+     * @throws IOException
+     */
     @Override
-    public int serializ(CRDT m) throws IOException {
+    public long serializ(CRDT m) throws IOException {
         return gen((SimpleNode<String>) m.lookup()).length();
     }
 }
