+++ src/main/java/crdt/tree/orderedtree/renderer/SizeJSonStyleDoc.java
@@ -72,8 +72,14 @@ public class SizeJSonStyleDoc implements SizeCalculator {
         return "{" + json(in) + "}";
     }
 
+    /**
+     *
+     * @param m the value of m
+     * @return the long
+     * @throws IOException
+     */
     @Override
-    public int serializ(CRDT m) throws IOException {
+    public long serializ(CRDT m) throws IOException {
         String xml = view((OrderedNode<String>) m.lookup());
         return xml.length();
     }
